Dict class was moved to their own .js file
