+++ index.html
@@ -12,8 +12,8 @@
   <script src="js/src/rrDisplay.js"></script>
   <script src="js/src/Claim.js"></script>
 
-  <script src="js/src/score.js"></script-->
-  <script src="js/src/SettleIt.js"></script>
+  <script src="js/src/score.js"></script>
+  <script src="js/src/SettleIt.js"></script-->
   <script src="js/src/Root.js"></script>
 
   <script src="bundle.js"></script>
