+++ js/src/Claim.js
@@ -1,4 +1,4 @@
-class Claim {
+module.exports = class Claim {
     constructor(id, isProMain) {
         /** The text of the claim with the claim. May include markdown in the future. */
         this.content = "New Claim";
