+++ js/src/Dict.js
@@ -1,0 +1,3 @@
+module.exports = class Dict {
+
+}
